Implement computer choice game logic
The game is 2-player and to not need another person,
an in-game computer is needed to go against the
player.

Prior to this commit, there wasn't an in-game computer.

Currently, I'm not familiar with arrays so I decided
to use Math.random and Math.floor to come up with
a random integer, that decide the computer's choice
against the player.