Implement human choice game logic
There can't be a game without the human player
who initiates it! Human choice game logic has now
been added.

Prior to this commit, there was no way to prompt
the player for their choice.

The approach I took prompts the player for their
choice, stores it, and simply returns it. The prompt
is strict though, in that it only takes the player's
choice if it has an uppercase first letter.

"Working okay is better than not working at all."

I'll definitely be changing this at a later date
though.