Implement play round game logic
Games don't go on forever - someone has to win and someone has to lose!
Rounds and increasing scores help determine who's winning and when one
game ends and when a new game begins.

The game had no round-based system or score increasing system before
this commit.

I went ahead and used switch cases to handle the logic in this section
of the game as it seems to handle the game's needs quite effectively and
is better than having a bunch of if..else conditionals for each outcome.

And in order to make the game's code neater, I renamed, added, cut out,
and shuffled around bits of code for better readability and to reduce
scope as to not evoke certain functions more than once.