Implement playGame() first to 5 points win limit
The game needs an artificial stop limit so players don't end up playing
forever!

With how I previously had the code, a game would only be 1 round as
playRound() was only evoked once. No previous commit intentionally imposed
a limit to how many rounds each individual game could be.

I decided to use a for loop that's based off of a max number of rounds
(20) combined with an if...else conditional (score checker) to implement a
score-based round system that sets a max number of 5 points as the imposed
limit for each game - with 20 rounds providing leeway until 5 points
are reached by one of the players.

And since the system goes off of the round and maxRound variables in the
for loop - the scores incremented within playRound(), the variables humanScore
and computerScore - can be incremented without interfering with the amount
of times that the for loop repeats.

Scores are now displayed to make keeping track of who's winning a
lot easier and I changed the way humanScore and computerScore incremented
as it did not work as intended.

Now, 2 bugs currently exist at the moment:

Bug #1 - both humanSelection and computerSelection stay the same throughout
the entire game. Meaning that after you input your choice for the first
time, it repeats that same choice for the entirety of the game. Same
happens with the computer's choice.

Bug #2 - on the last round of the game, after the score reaches 5, the
prompt that takes the human player's input shows up one last time before
the score checker if...else conditional logs the actual winner of the game.