Refactor both in-game player's logic
The previous version of the game was too long and clunky.

Starting with both of the player's getting functions - they've have rewritten
to shorten overall code length and better fit the need of the new code's structure.

I decided to make GetComputerChoice() use an array tied in with a random number generator
that then serves to return one of the 3 choices as it considerably reduces the lines of code
needed.

Both functions now also return lowercased strings as they better fit the overall needs of the
game later down the line.