Refactor human player input method
To use buttons instead of a prompt to input the human player's choice, a
few changes were in order.

With how I previously coded the game, the human player could only type
in their option instead of just clicking the button with their option already hardcoded in
so that game could run.

I ended up hardcoding the button into the HTML file since it's one of my first times using the DOM
and it would be easier for me to understand it if I didn't have to have a constant mental model.

I also - for simplicity sake - deleted the for loop that served as a way to play each round until the winning score was hit.

And note - the game currently isn't working as intended.