Refactor playGame() and playRound()
PlayGame() and playRound() both needed to be refactored to fit the new
requirements of the code.

playGame() now encapsulates all of the currently existing code. I did this
because the buttons that now serve as the human player's way of inputting
their choice needs access to playRound() to work properly.

playRound() now also uses the DOM to add the appropriate text to the .display-area div.

Note - the game still does not work as intended.