Add in-game score checker - game now works!
Without a proper way to determine the winner, the game goes on forever...

Now, the game stops when one of the players reaches the winning score.

I did have a score checker before but with how I had implemented it, it was outside of the playRound()
function and relied on the variable maxRounds and a for loop to work properly.

With how I have it now, the score checker is implemented with playRound(), so every time a button
is clicked, only the function playRound needs to runs.